requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
python-dotenv==1.0.0
deepgram-sdk==3.2.0
groq==0.4.2
requests==2.31.0
